Return the public address of a peripheral that has just been paired.
A random address may be used when pairing to a peripheral. For later reference, the public
address needs to be known. Until this change, there was no way to know the public address
from the information available when performing the pairing. The kernel returns the
public address at pair completion, and bluepy now propogates it to the caller.